+++ js/examples/react_example/src/Animation.js
@@ -1,11 +1,11 @@
-import React, { useRef, useEffect, useState, FunctionComponent } from 'react';
-import { RiveAnimation } from 'rive-js';
+import React, { useRef, useEffect } from 'react';
+import { Rive, Layout } from 'rive-js';
 import styles from './App.css';
 
 const Animation = ({ asset, animation, fit, alignment }) => {
     const canvas = useRef(null);
     const animationContainer = useRef(null);
-    let riveAnimation;
+    let rive;
 
     // Resizes the canvas to match the parent element
     useEffect(() => {
@@ -16,14 +16,15 @@ const Animation = ({ asset, animation, fit, alignment }) => {
 
     // Start the animation
     useEffect(() => {
-        riveAnimation = new RiveAnimation({
+        rive = new Rive({
             src: asset,
             canvas: canvas.current,
-            // alignment: new CanvasAlignment({ fit: fit, alignment: alignment }),
+            animation: animation,
+            layout: new Layout(fit, alignment),
             autoplay: true,
         });
 
-        return () => riveAnimation?.stop();
+        return () => rive?.stop();
     }, []);
 
     return (