+++ js/src/rive.test.ts
@@ -1,5 +1,0 @@
-// const { testables } = require('./rive.js');
-
-test('Tests go here', () : void => {
-  expect(true).toBeTruthy();
-});